Changed to ray weighting scheme.

+++ src/sim/render/painter/mod.rs
@@ -12,5 +12,4 @@ use palette::LinSrgba;
 use rand::rngs::ThreadRng;
 
 /// Fragment painter function.
-pub type Painter =
-    fn(&Point3<f64>, &Grid, &Shader, &Scheme, Ray, &mut ThreadRng, usize) -> LinSrgba;
+pub type Painter = fn(&Point3<f64>, &Grid, &Shader, &Scheme, Ray, &mut ThreadRng, f64) -> LinSrgba;
